feat ( Adiciona event de login e arruma exclusão de usuário: - Adicionando evento no login - Adicionando listener do evento de login - Adicionando registro do log de login no banco - Adicionando tratamento no método de excluir usuário impedindo auto exclusão e desabilitando ícone da lixeira no front)

+++ src/services/Aluno/AlunoService.ts
@@ -171,7 +171,12 @@ class AlunosService{
         return alunoAtualizado;
     }
 
-    async delete(id : string) : Promise <Aluno>{
+    async delete(id : string, usuarioId : string) : Promise <Aluno>{
+        
+        if (id === usuarioId) {
+            throw new Error ("AUTO_EXCLUSAO_NAO_PERMITIDA")
+        }
+
         const alunoDados = await this._alunoPrismaRepository.delete(id)
 
         if (!alunoDados){
